requirements.txt
tensorflow
keras
pillow
scipy